WEBAPP-126: Move LocationFooter and LocationHeader to components
