+++ src/modules/app/components/LocationFooter.js
